Разбил проект на части
